# Run unit tests in parallel by default. 