--- a/lib/coin-selection/cardano-coin-selection.cabal
+++ b/lib/coin-selection/cardano-coin-selection.cabal
@@ -99,3 +99,4 @@ test-suite test
     Cardano.CoinSelection.CollateralSpec
     Cardano.CoinSelectionSpec
     Spec
+    SpecHook