--- a/lib/coin-selection/test/spec/SpecHook.hs
+++ b/lib/coin-selection/test/spec/SpecHook.hs
@@ -1,0 +1,6 @@
+module SpecHook where
+
+import Test.Hspec
+
+hook :: Spec -> Spec
+hook = parallel