--- a/lib/primitive/cardano-wallet-primitive.cabal
+++ b/lib/primitive/cardano-wallet-primitive.cabal
@@ -166,3 +166,4 @@ test-suite test
     Cardano.Wallet.Primitive.Types.UTxOSpec
     Data.QuantitySpec
     Spec
+    SpecHook