--- a/lib/primitive/test/spec/SpecHook.hs
+++ b/lib/primitive/test/spec/SpecHook.hs
@@ -1,0 +1,6 @@
+module SpecHook where
+
+import Test.Hspec
+
+hook :: Spec -> Spec
+hook = parallel