--- a/lib/strict-non-empty-containers/strict-non-empty-containers.cabal
+++ b/lib/strict-non-empty-containers/strict-non-empty-containers.cabal
@@ -70,3 +70,4 @@ test-suite unit
       Main.hs
   other-modules:
       Data.Map.Strict.NonEmptyMapSpec
+      SpecHook