… lib/primitive/test/spec/Cardano/Wallet/Primitive/Types/TokenQuantitySpec.hs
… lib/primitive/test/spec/Cardano/Wallet/Primitive/Types/UTxOIndexSpec.hs
… lib/primitive/test/spec/Cardano/Wallet/Primitive/Types/UTxOSelectionSpec.hs
… lib/primitive/test/spec/Data/QuantitySpec.hs
… lib/strict-non-empty-containers/test/unit/Data/Map/Strict/NonEmptyMapSpec.hs
… lib/strict-non-empty-containers/test/unit/SpecHook.hs
… lib/wallet/test/unit/Cardano/Byron/Codec/CborSpec.hs
… lib/wallet/test/unit/Cardano/Pool/DB/SqliteSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Address/PoolSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Api/TypesSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/ApiSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/DB/LayerSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/DelegationSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDerivation/ByronSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDerivation/IcarusSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDerivation/MintBurnSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDerivationSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDiscovery/RandomSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDiscovery/SequentialSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDiscovery/SharedSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/AddressDiscoverySpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/BlockSummarySpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/Migration/PlanningSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/Migration/SelectionSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/MigrationSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/ModelSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/Passphrase/LegacySpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/PassphraseSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/SlottingSpec.hs
… lib/wallet/test/unit/Cardano/Wallet/Primitive/SyncProgressSpec.hs