This PR adjusts our unit test suites to run tests in parallel by default, making it unnecessary to manually use the  `parallel` combinator for every individual `Spec`.

For the `cardano-wallet:test:unit` test suite in particular, this leads to a large speedup when run on a machine with at least 4 cores:

### Original timings
```
Finished in 705.3929 seconds, used 1913.9534 seconds of CPU time
real    11m49.908s
user    25m45.315s
sys     6m17.046s
```

### After using `SpecHook` to run tests in parallel by default
```
Finished in 500.0908 seconds, used 1755.7065 seconds of CPU time
real    8m25.017s
user    25m7.003s
sys     4m17.505s
```

### After removing all manual calls to `parallel`
```
Finished in 506.2987 seconds, used 1779.8373 seconds of CPU time
real    8m31.305s
user    25m23.259s
sys     4m25.264s
```