@jonathanknowles 2023-04-14
I noticed the following failure:

```
cardano-wallet-test-unit> Failures:
cardano-wallet-test-unit>   test/unit/Cardano/Wallet/Shelley/NetworkSpec.hs:196:9:
cardano-wallet-test-unit>   1) Cardano.Wallet.Shelley.Network.Observer, typical use, when refresh fails, only MsgWillFetch is traced
cardano-wallet-test-unit>        expected: [MsgWillFetch (fromList ["k"])]
cardano-wallet-test-unit>         but got: [MsgWillFetch (fromList ["k"]),MsgRemovedObserver "k",MsgWillFetch (fromList []),MsgDidFetch (fromList [])]
cardano-wallet-test-unit>   To rerun use: --match "/Cardano.Wallet.Shelley.Network/Observer/typical use/when refresh fails/only MsgWillFetch is traced/"
```